working bcm2836sdhc in arm64, still rough

+++ drivers/sd/bcm2836/bcm2836sdhc/bcm2836sdhc.c
@@ -2500,7 +2500,7 @@ VOID
 SdhcReadDataPort (
     PSDHC_EXTENSION SdhcExtension,
     PUCHAR Buffer,
-    ULONG Length
+    SIZE_T Length
     )
 {
     ULONG ByteCount = Length % sizeof(ULONG);
